[Feat] Lotto Func Move to Service
