+++ app/src/main/AndroidManifest.xml
@@ -20,21 +20,16 @@
         <activity
             android:name=".presentation.chatting.SocketChattingActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
 
         </activity>
         <activity
             android:name=".presentation.coroutine.WorkManagerActivity"
             android:exported="true" />
-<!--
-        <service
-            android:name=".MusicPlayerService"
-            android:enabled="true"
-            android:exported="true" />-->
+        <!--
+                <service
+                    android:name=".MusicPlayerService"
+                    android:enabled="true"
+                    android:exported="true" />-->
 
         <activity
             android:name=".presentation.musicplayer.MusicPlayerActivity"
@@ -44,7 +39,13 @@
             android:exported="true" />
         <activity
             android:name=".presentation.coroutine.CoroutineLottoActivity"
-            android:exported="false" />
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".presentation.coroutine.CoroutineExampleActivity"
             android:exported="false" />
@@ -67,17 +68,17 @@
             android:name=".presentation.broad.BroadcastPracticeActivity"
             android:enabled="true"
             android:exported="true" />
-<!--
-        <receiver
-            android:name=".SMSReceiver"
-            android:enabled="true"
-            android:exported="true"
-            android:permission="android.permission.RECEIVE_SMS">
-            <intent-filter>
-                <action android:name="com.sopt.androidstudy" />
-                <action android:name="android.provider.Telephony.SMS_RECEIVED" />
-            </intent-filter>
-        </receiver>-->
+        <!--
+                <receiver
+                    android:name=".SMSReceiver"
+                    android:enabled="true"
+                    android:exported="true"
+                    android:permission="android.permission.RECEIVE_SMS">
+                    <intent-filter>
+                        <action android:name="com.sopt.androidstudy" />
+                        <action android:name="android.provider.Telephony.SMS_RECEIVED" />
+                    </intent-filter>
+                </receiver>-->
     </application>
 
 </manifest>
