+++ app/src/main/java/com/sopt/androidstudy/data/remote/lotto/LottoService.kt
@@ -1,0 +1,7 @@
+package com.sopt.androidstudy.data.remote.lotto
+
+import java.util.ArrayList
+
+interface LottoService {
+    suspend fun getLottoNumbers(round:Int): ArrayList<Int>
+}
