+++ app/src/main/java/com/sopt/androidstudy/presentation/coroutine/CoroutineLottoActivity.kt
@@ -83,31 +83,4 @@ class CoroutineLottoActivity : AppCompatActivity() {
             ivGame5.setImageResource(getDrawableID(result[5]))
         }
     }
-
-    private suspend fun getLottoNumbers(): ArrayList<Int> {
-        val round = "1000"
-        val url = "https://dhlottery.co.kr/common.do?method=getLottoNumber&drwNo=$round"
-        val lottoNumbers = ArrayList<Int>()
-        try {
-            withContext(Dispatchers.IO) {
-                val response = URL(url).readText()
-                val jsonObject = JsonParser.parseString(response).asJsonObject
-                val returnValue = jsonObject.get("returnValue").asString
-                if (returnValue == "success") {
-                    for (i in 1..6) {
-                        val lottoNumber = jsonObject.get("drwtNo$i").asInt
-                        lottoNumbers.add(lottoNumber)
-                    }
-                    val bonusNumber = jsonObject.get("bnusNo").asInt
-                    lottoNumbers.add(bonusNumber)
-                    lottoNumbers.add(round.toInt())
-                }
-            }
-        } catch (e: Exception) {
-            e.printStackTrace()
-        }
-
-        return lottoNumbers
-    }
-
 }
